backend/app/requirements.txt
alembic==1.8.1
redis>=4.5.0
Flask-Caching==2.1.0
Flask-Session==0.4.0
Flask-Limiter==3.3.1
aniso8601==9.0.1
appdirs==1.4.3
APScheduler==3.10.4
attrs==25.4.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
distlib==0.3.6
distro==1.4.0
filelock==3.8.0
flasgger==0.9.7.1
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
flask-marshmallow==0.14.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
greenlet==3.3.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lockfile==0.12.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.29.0
mistune==3.2.0
msgpack==0.6.2
packaging==21.3
passlib==1.7.4
pipenv==2022.11.5
platformdirs==2.5.3
progress==1.5
psycopg==3.1.4
psycopg2-binary==2.9.11
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytoml==0.1.21
pytz==2022.1
PyYAML==6.0.3
qrcode==6.1
referencing==0.37.0
retrying==1.3.3
rpds-py==0.30.0
setuptools==82.0.0
short_url==1.2.2
six==1.14.0
SQLAlchemy==1.4.54
toml==0.10.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==1.25.8
virtualenv==20.16.6
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.0.3
zipp==3.10.0
